Stop URL-encoding pub/sub topics
We had this grand idea in the past when adding the `publish` and
`subscribe` operators to make topics hierarchical. Because of that, we
URL-encoded them, envisioning use of the forward slash as a separator.
In practice, what we've ended up doing is really treating the topics as
totally separate channels, and using schema names to form a hierarchy if
we want one.

Unforutnately, the URL-encoding became user-visible with the
introduction of the `publish` and `subscribe` metrics, which contained
the URL-encoded topic in metrics. This was confusing not just for users,
but also for ourselves in production setups we manage.

This change removes the URL-encoding entirely. It also changes the
default topic to be `main`.