Allow aggregation functions to be saved and restored
This is in preparation for lookup-table contexts with explicitly
specified fields via aggregation functions, which need to be able to
roundtrip correctly.

I've split this out into a separate PR to make this easier to review;
tests for this will arrive with the follow-up story that makes the
changes to the contexts, as only then this newly added code will be
reachable by users.

Note that `median` and `quantile` cannot be serialized and deserialized
as of this PR, as they internally use Apache Arrow's TDigest
implementation, which is unfortunately hidden via the pimpl idiom. We
will need to change that in a follow-up PR to use another TDigest
implementation that we can actually serialize.