Mask secrets in `/serve`
This adjusts the `serve` operator to replace all secrets as defined in
the `tenzir.secrets` configuration section for the `secret` function
with `***`. This effectively causes secrets not to show up in the
Tenzir Platform anymore.

Note, though, that this is not yet a complete solution, as the download
button starts a pipeline that does not go through the `serve` operator,
but rather uploads events to the platform's S3-compatible storage so
that the user's web browser can download the file directly afterwards.
To fix this, we would need to adjust how that data is uploaded.

I'm not certain whether we want to merge this as-is, or whether we want
to consider alternative approaches in the future.