Flatten kafka options from the config files
Tenzirs internal logic unflattens all config options with a '.' in
a key. This is generally not a great idea, but that fix has to wait
till another time, as it will be very far reaching.

For now we just flatten all options that are being passed to rdkafka
again to make the kafka plugin work.