Fix bugs in `where` when predicate evaluates to `null` (#4785)
This fixes two bugs in the `where` operator when the evaluated predicate
contained `null` values:
1. When the result only contained `true` and `null`, we incorrectly
returned the entire input.
2. When the result contained `null` and at least one `false`, we ran
into UB by incorrectly accessing the value of an `arrow::BooleanArray`
for an invalid row.

Now, we correctly warn about `null` values in the result and drop all
events for which the predicate evaluated to `null`.