Add copy button to named language code blocks
Once again, there were quite a few hoops to pass through.

With the `jsx: true` option in the transformer, it disables the
onclick handler on the transformer, which we manually add later
on each page that contains the correct class name. See the
implementation of the `registerCopyButton` function to understand
what's going on here.

I expected docusaurus component swizzling to work, but adding the
registerCopyButton as a hook (as the `@rehype-pretty/transformers`
library suggests) to /theme/Root.tsx or /theme/Layout/index.tsx
didn't do the job.