Allow aggregation functions to be called on lists
This enables `[1, 2, 3].sum()` and all the other combinations you can
think of. This was already documented, but not implemented.

The current implementation is done by mapping the list array's value
array onto a table slice with a single field called `x`, and swapping
the first argument of the aggregation function call with a root field
called `x`. That's a bit hacky, but the alternative would've required
changing the interface of aggregation functions not to evaluate their
arguments on table slices directly anymore. Note that this effectively
adds the restriction that the first argument of an aggregation function
must be the aggregation function's subject, which holds true for all
aggregation functions we currently have.