Fix a rare crash in the index actor on startup
On startup, the node actor starts its core components in order of how
they depend on another. However, with the addition of the
`tenzir.ingest` metrics, which are measured in the index component, we
introduced a circular dependency between the importer and index
components.

This change fixes a rare crash when the importer was started after the
index tried emitting its first metrics. We observed this a few times in
CI, and I have never heard of this happening in production.