Fix GCS plugin for Arrow 19+
Arrow 19 introduced a breaking change for `GcsFileSystem::Make(…)`,
which now returns a `Result<shared_ptr<FileSystem>>`, despite not having
a single failure path in its current implementation, with upstream's
rationale[^1] being consistency with the other filesystem
implementations.

Additionally, this bumps the Arrow version used in the Docker images to
from v18.1.0 to v19.0.0.

[^1]: https://github.com/apache/arrow/pull/44503