Avoid compiling twice in the static binary packaging
The way our deb/rmp/tarball packaging was set up caused a recompilation
of all source files after the regular installation of the runtime
files into the Nix store.
The reason for that is that we had to reconfigure the build tree
to set the correct package installation directory for CPack, and
CMake incorrectly detected the the object files needed a rebuild.

To work around that issue, we now simply invert the order and build
the packages directly in the buildPhase and unpack the tarball into
$out for the installation.