Add `fill` option to chart operators
This option replaces all `null`s in the output of the operators with the
specified value. The type of value must match `y_min` and `y_max` if present.

Since this option can never be `null`, we can now fill gaps in all charts, not
just `chart_line`, when specified.

Currently, `resolution` must be specified alongside, which means this cannot be
used with `chart_pie`. `chart_pie` does not have `resolution` exposed as it has
a `label` option rather than an `x` option. If we have a better interface/name
in the future, we could expose all options.

Additionally, now:
- All buckets in a gap are filled, not just the first one.
- `x_min` and `x_max` must hold valid types for X-Axis.
- `y_min` and `y_max` must hold valid types for Y-Axis.
- y-attributes are sorted.