Distinguish between Blogs and Releases on docs site
Now that we have the new blog up on tenzir.com/blog, it makes sense to
distinguish here. Historically, our blog on the docs site was mostly
used for release announcement anyways, so this isn't too bug of a
change.

I left the URL at docs.tenzir.com/blog as I couldn't find a way to do a
wildcard redirect to docs.tenzir.com/releases to avoid breaking existing
links to it.