Fix out-of-process REST APIs
This fixes a regression introduced in a recent change I made that
removed usage of the node actor's message handler to retrieve component
handles in favor of using the regisrty where possible.

It turns out that we still have the possibility to spawn the REST API
externally via `tenzir-ctl web server`, which under-the-hood spawns
`rest_handler_actor` instances in a client process. After my recent
change this now ran into an assertion failure, as I had naïvely assumed
that the REST handlers were singletons in the node.

This change now reverts the recent changes for the two affected actors.
A proper fix would be to change the actors to run inside the node, but
as we do not support building the `web` plugin locally on macOS I am
likely not the right person to be making that change.

To test this change, run `TENZIR_VERSION=topic-fix-external-api docker
compose up` from the repository root after CI has finished uploading the
image for this PR. If the `tenzir-api` service starts instead of running
into an assertion failure (like it does on current main), then this
change works.

No changelog entry as I introduced this regression only after the last
release.