Do not expose the CAF port in Docker Compose
We have `docker compose run tenzir` to interact with the Tenzir Node
from within the Docker network already, so there's no reason to expose
the port to the outside world.

This makes to easier to run `docker compose up` multiple times on the
same host.

In addition, this makes the REST API opt-in via `docker compose run
tenzir-api`. This does need to expose the port to the outside world, and
there was little reason to always enable it as it is rarely used.