Remove type layering
This is a bug fix with unfortunate consequences.

Previously, assigning metadata like attributes or names to types
could—depending on the oredr in which metadata was asigned—change the
physical layout of the type's underlying FlatBuffers table. This, in
return, caused types that were logically equivalent not to compare equal
and the `type_id` function in TQL2 to return unexpected values.

Here's an easy reproduction:

```
from {x: 1, y: 2}
chart_line x=x, y=y
// Uncomment this line to change the resulting `id`:
// x = x
id = type_id(this)
```

It is clearly a bug that an assignment like `x = x` for data that
contains a field `x` changes the result of the `type_id` function.

This change removes the layering of types entirely, making the order in
which metadata is assigned irrelevant. This changes the `type_id` for
types with metadata now once, but makes it more stable in the future.

Unfortunately, we have in the past decided to store `type_id(this)` for
dashboard widgets in the Tenzir Platform. We stopped doing this for
Tenzir Platform v1.7, but any table or chart added to a dashboard before
then is likely to break, and will need to be re-added to the dashboard.
Newer dashboard widgets are not affected. We are terribly sorry about
this inconvenience, but did not find a good way to fix this.