Allow assigning ids for pipelines manually
The web is a lossy place. If a request from the Tenzir Platform to
deploy a pipeline does not get a response, we ideally want to retry
it—but that could potentially have bad side effects like deploying a
pipeline twice if sending the same request again can create a different
state than the one we wanted to have. Ultimately, the platform cannot
know if the node received a request when it did not get a response.

This changes the `/pipeline/create` and `/pipeline/launch` endpoints to
optionally accept a new `id` parameter for manually assigning a pipeline
id. That's far from a perfect fix as duplicate requests will still get
responded to with an error, but it's the best we can do for now without
majorly revisiting the code behind the `/pipeline` API.