Fix a crash in the MSB in merging mode with schema
If the MSB is running in merging mode, it directly acts on a series_builder.
However, if a schema is given, and we try to insert a list or record,
where the schema does not expect one, we run into an assertion failure.
This change performs a check first and avoids writing at all. The
series_builder will then implicitly null the field for us.