Fix mismatch in type metadata after assignments
This fixes a crash in Debug builds (via `TENZIR_ASSERT_EXPENSIVE`) in
the pipeline `from {} | x = this`, or, more generally speaking, any
pipeline that involves an assignment on a field which has metadata.

The expensive assertion that fired here compared that Tenzir's and
Arrow's metadata on Tenzir's types and Arrow's fields and schemas
matches exactly. A performance optimization made in Tenzir Node v4.28
introduced this bug that under some circumstances caused a mismatch
between the two.

The impact of this bug is luckily only small. Events persisted with
`write_feather`, `write_parquet`, or `import` restore the metadata from
Arrow when read. This implies a loss of metadata for nested fields with
metadata persisted since the release of v4.28, but luckily the data is
not in a broken state on disk.