Turn tenzir-plugins fetching into a fixed output derivation
This should make it possible for users on Nix to get the community
edition package from our cachix cache, because the private repo is
not required at evaluation time any more.

With this PR we also don't check if the submodule and the fetcher
source are in sync any more. Instead the Nix CI workflow  will do it
in-place.