Change how signed ints are rendered in logs
This removes the leading plus for positive signed integers in
`tenzir::data` when rendered as a string.

This came up as part of a bug search in the `from_fluent_bit` and
`to_fluent_bit` operators, where `options={port: 9200}` got forwarded as
`port=+9200` to Fluent Bit, and was then just ignored because of that
pesky little plus character.

The reason for why this leading plus existed is long gone: `vast export
ascii <predicate>` hasn't been a thing in years, and that relied on our
internal data printer (and parser) for rendering, introducing a need to
distinguish between signed and unsigned integers.