Fix cache eviction always happening on maximally large caches
When a cache would grow beyond the byte size configured in
`tenzir.cache.capacity`, a pipeline warning is emitted, and the cache operator
no longer accepts new events. Offending events are discarded, meaning that the
cache will have stopped just short of exceeding the limit.

Additionally, caches all report their sizes up to a central cache eviction
service which periodically evicts older caches that cause the global limit to be
exceeded.

However, currently the reported size of the a cache that "stopped just short" of
exceeding the limit includes the size of the events that would've caused it to
be exceeded, leading the eviction service to immediately evict the cache.

This commit makes it so the size of offending events isn't included in the size
reported to the eviction service.