Implement and use common component for SSL control
This makes use of the new common `ssl_options` in every operator that has
TLS options. TLS is now defaulted to `true`, except in TQL1 or TCP.
The cacert file is picked from the global config, unless its explicitly
set on an operator. If its not set in the tenzir.yaml, we try to find
the default cacert for the system.