Fix hang in `cache` when creating an empty cache
This fixes an infinite hang in `<source> | head 0 | cache "<id>"`, which
successfully created the cache, but then failed to shut it down. This
happened because the pipeline writing into the cache lazily registered
itself at the cache upon writing the first batch of events. The fix was
to simply register beforehand.