Move rebatching into the `importer` actor
Previously, we've rebatched events on import in the `import` operator.
This, however, had a significant flaw: It did not work for events sent
to the `importer` actor directly, e.g., metrics and diagnostics. This
caused access to recent metrics and diagnostics to be very expensive due
to the smaller batch sizes.