Add 'libcap2-bin' dependency to Dockerfile
This allows the usage of the 'capsh' binary in our docker
compose setups. The capsh binary in turn is currently the
easiest way to manipulate ambient capabilities from the command
line, which are required to allow the tenzir node to e.g. bind
to privileged ports in host networking mode.