Add first class `secret` type.
This adds a first class type `secret` to Tenzir's type system. The type
contains of a secret value, a name for the secret (used when fetching)
from a secret manager, as well as some other information about it, used
when accessing or persisting it.

A secret's value is not supposed to persisted on disk (via import or
checkpointing) and should not be send anywhere via operators/functions.
This needs to be ensured in the C++ codebase. To easily strip secret
values, a C++ function is provided.

Additionally, this PR adds simple functions to create proper secrets
from strings and query ones from the config. Secret-Manager integration
will be in a followup PR. A hidden function to get the plain value out
for testing is also provided, but only works in debug mode.