Hardcode the fluent-bit page size to 24576
This is larger then the page size of all currently supported
platforms, with apple silicon being the biggest at 16K.

The specific value 24576 is coming from the fluent-bit configuration
docs, claiming it to be the default value. The code must have gone
out of sync from the docs, so we now set it explicitly.

This change fixes a segfault when the fluent-bit engine is rebatching
large-ish events in its internal format. For example

    plugins
    repeat 1000
    to_fluent_bit "http", fluent_bit_options={
        host: "localhost",
        format: "json_lines"
    }

would reliably crash before this change.