Make TQL2 the default
This change makes TQL2 the default (and the only option) by removing the
`TENZIR_TQL2` environment variable and the corresponding `tenzir.tql2`
option and `--tql2` command-line flag.

Note that this must rewrite all old BATS tests from TQL1 to TQL2; turns
out we have a lot more of these than I had thought, but it's well worth
the effort.

The next node release after this PR is in will be Tenzir Node v5.0.