Remove 20k element limits from `data_builder`
These arbitrary limits were put there during development to prevent
unbounded memory consumption and retention by the `data_builder`, as its
internal structure would be kept around (for both optimization and
batching reasons) for the entirety of a pipeline.

With this change, the limit is gone, and instead the memory is released
after every event, if it exceeded the size limit.