Fix a stack-use-after-move in `save_tcp`
It's not save to issue a request from instantiate from the legacy saver
API and using exec-node state in the corresponding response hnadler.

This rewrites `save_tcp` with the operator API, which as a byproduct
fixes this problem, as we can now properly yield control to the exec
node after issuing the request.