Experiment with a trailing `?` for field access
Now that we're working more with suppressed warnings in field accesses,
I wanted to experiment with a trailing `?` instead of using a separate
`.?` operator for field access:
- `this.?foo` is now `foo?`
- `foo.?bar` is now `foo.bar?`
- `foo.get(bar, null)` is now `foo[bar]?`

Just opening this for comments; not sure if we want to go forward with
this. This came up because of the recent work on the `move` keyword,
where moving top-level fields becomes a lot more readable with `move
foo?` instead of `move this.?foo`.