Introduce lambda functions
This adds lambda functions to TQL, used as arguments for the `map` and
`where` functions, and additionally for the newly added `count_where`
aggregation function.

I'm opening this PR to discuss the feature before I follow up on it—as
of now this is an hour of hacking on it over the weekend, and nothing
more. There's a fair amount of work to be done on this before we could
merge it.