Always shut down the pipeline manager first
This fixes a rare race condition on shutdown that was observed by a user
and reported via Discord. On shutdown, the pipeline manager component,
which, as the name suggests, manages the lifetime of running pipelines
in the node, needs to take special care that it is the instance that's
shutting down pipelines, so that it knows not to update the pipelines'
last known state to `stopped`, `completed`, or `failed`.

This change fixes a bug that could occur due to an indeterministic
shutdown order, where if another component shut down first it could
consequentially cause dependent pipelines to complete or fail. The
pipeline manager would then record these new states before shutting down
by itself, and the pipelines would not get restarted on startup.

The bug was pretty rare, and more likely to happen under high load. In
my own testing, I could observe it just three times in over 10k
restart cycles. I am confident that this fixes the bug after going
through another 10k cycles, but it is hard to tell exactly, and
practically impossible to write a test for this in our existing test
framework.