Do not refuse startup when `pid.lock` is invalid
During my many tests of for tenzir/tenzir#5163, I sometimes stopped the
node after the `pid.lock` file was created but before it was written to,
causing it to be empty. Afterwards, my node always refused to start up
again until I manually deleted the `pid.lock` file.

This change makes it so that we warn the user about the incorrect
shutdown, but still start up, overriding the file.