Change `secret` flatbuffer and treat secrets as bytes.
`secret`s now only contain a flatbuffer, an their arrow arrays are binary
arrays. This makes the type forward compatible.

We no longer assume that resolved secrets are utf-8 strings.
Instead, the assumption now is that they are bytes. A convenience function
to try and get the secret value as utf-8 is provided, as that is all of
our current usecases.

Finally, this implements secret concatenation and decode/encode using the
new flatbuffer backend.