Experiment with making the run loop of exec nodes cheaper
Not sure if this is going to pass all tests yet, this is a bit of an
experiment. This does change the shutdown behavior of exec nodes ever so
slightly, as they now keep themselves alive. Consequently, if any of
them relied on becoming unreachable for shutting down we will now have a
hang. Let's see what happens.