Add an experimental `integration-coverage` target
This generates a coverage report from our integration tests inside. It
doesn't run any other tests, so the numbers do look a lot worse than
they should.

The goal for this is not to write tests for improved coverage. It is for
developers to be able to easily identify which branches of newly added
code are not covered by tests.

I have only tested this on macOS with LLVM 20.1.5.

There is no integration with CI for this yet in any way, however I do
think that this is safe to merge as developer-facing tooling.