Fix the backwards-compat mode for lambdas
`[1, 2, 3].map(x, x + 1)` should emit a warning and function just the
same as `[1, 2, 3].map(x => x + 1)`. Unfortunately, during a rebase, I
must've accidentally removed the line that assigned the left-hand side
of the lambda expression in this backwards-compat mode.

No changelog entry is needed for this one, as the lambda change hasn't
been released yet.