Fix serializer roundtrips of empty non-null chunks
This fixes a bug that caused `context::create_geoip "name"` to fail when
no `db_path` option was provided. A recent refactoring in CAF caused a
subtle behavior change for CAF deserializer and serializer
implementations that used the old mechanism from up to CAF 0.17 instead
of the newer one introduced with CAF 0.18.