Add `/serve-multi` endpoint
This endpoint works similar to `/serve`, but accepts an array of
`requests: [{serve_id:string,continuation_token:string}]`. The reply
will be a record keyed by the serve IDs, with the each of the values
having the same form as from `/serve`.

The min&max events query parameters are evenly distributed across all
serve_ids.

The node provides a feature flag `serve-multi` to indicate this feature
is available.