Remove type name form the OCSF enrichment column
This avoids an error with the Tenzir and Arrow type differing. The
underlying issue is not fixed.

We are still failing to set some Arrow metadata in `set.cpp::assign` and
the `series_builder` does not seem to build any arrow metadata at all.

It would appear that its mostly OK, the "only issue" is an assertion in
`make_table_slice` that is not enabled in release, but of course provides
an important check.