Handle null `time` in `to_amazon_security_lake`
Previously events with a null time would incorrectly be added to the last
group. This is both incorrect and could cause a crash. These events are
now correctly skipped.

Additionally, the operator would previously not relinquish control when
skipping an entire slice, which would mean it would be stuck waiting for
input.