Simplify nix workflow execution
We now always run it and let Nix figure out whether a rebuild is
necessary. No need to come up with manual heuristics while Nix can
already do this properly.

This also replaces the nushell script for the dynamic parts of the
build with a python script, and integrates registry login code into
that.